Separate stateful and stateless parts
